requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.17.0
Flask==3.0.3
flask-cors==5.0.1
frozenlist==1.5.0
fsspec==2025.3.0
gotrue==2.11.4
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
itsdangerous==2.1.2
Jinja2==3.1.6
jiter==0.9.0
llvmlite==0.43.0
lxml==5.3.1
MarkupSafe==3.0.2
more-itertools==10.6.0
moviepy==2.1.2
mpmath==1.3.0
multidict==6.2.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
openai==1.66.3
openai-whisper==20240930
packaging==24.2
pillow==10.4.0
postgrest==0.19.3
proglog==0.1.10
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
PyMuPDF==1.25.3
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
realtime==2.4.1
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.14.0
supafunc==0.9.3
sympy==1.13.3
tiktoken==0.9.0
torch==2.2.2
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.2
Werkzeug==3.1.3
yarl==1.18.3
zipp==3.21.0
